requirements.txt
boto3==1.26.66
botocore==1.29.66
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.1
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
moto==4.1.2
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
responses==0.22.0
s3transfer==0.6.0
six==1.16.0
toml==0.10.2
types-toml==0.10.8.3
urllib3==1.26.14
Werkzeug==2.2.2
xmltodict==0.13.0
